create DOM manipulation folder to hold all dom manipulations functions
