+++ lib/swatches.js
@@ -1,6 +1,9 @@
 const $ = require('jquery')
 import COLORS from './data/colors'
 import swatchHelpers from './helpers/swatches_helpers'
+import colorAppender from './DOM_manipulation/swatches_appender'
+
+
 
 $('button').click(function(){
   createSwatches()
@@ -20,13 +23,4 @@ const createSwatches = () => {
   swatchHelpers.colorSifter(splitText, colorArray)
   const uniqueText = swatchHelpers.unique(colorArray)
   colorAppender(uniqueText)
-}
-
-
-const colorAppender = (arrayOfColors) => {
-  arrayOfColors.forEach(function(color){
-    let hexValue = COLORS[color]
-    $('article.colorized-text')
-    .append(`<div class="swatch" style="background-color:${hexValue};"></div>`)
-  })
 }
